detect flutter module types in small IDEs
